site.db gets stored in the proper data / database directory
I made some changes how the site.db gets stored in the right directory. It now gets stored in the data / database directory instead of the same place where main.py is located in.

I made two users as a test to check if the password is indeed another password. This checks out correctly.