Fixing issue with wrong prices at the shop
I don't know what the purpose of "// 2" was, but it leads to wrong prices in the shop (shop menu shows "$2" for corn, but it actually costs $4). Since I don't want to change the code, I replace "//2" with "//1" to get rid of this little problem. The code can be improved by someone else if we have the time.